requirements.txt
asgiref==3.4.1
Django==3.2.5
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-writable-nested==0.6.3
psycopg2==2.9.1
PyJWT==2.1.0
pytz==2021.1
sqlparse==0.4.1
views==0.3
